Add a custom mocha test runner for better console.log handling
Before we were forced to use console.warn to throw debug logs because
waiting for the start of the JSON output was a bit tricky. Back on
a stream-based parser (and one we fully control) we can be pretty
confident that a line is or is not a test status line. AND we get
to keep the diff!

Nice!

It'd be great to actually perform a diff and color code it and all,
but that'll take a bit more work...